pyproject.toml
[tool.poetry]
name = "timer"
version = "0.1.0"
description = ""
authors = ["Vladislav Marahovskiy <vladislav.marahovskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.2"
sqlalchemy = "^2.0.32"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.112.0"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
ruff = "^0.5.6"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
dependency-injector = "^4.41.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
